+++ src/main/java/com/github/choonchernlim/security/adfs/saml2/CustomAuthnContext.java
@@ -1,0 +1,11 @@
+package com.github.choonchernlim.security.adfs.saml2;
+
+/**
+ * ADFS-specific authentication method.
+ */
+public final class CustomAuthnContext {
+    /**
+     * Windows integration authentication.
+     */
+    public static final String WINDOWS_INTEGRATED_AUTHN_CTX = "urn:federation:authentication:windows";
+}